requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.3
pytest==7.4.4
httpx==0.26.0
markdown==3.7